Memory efficiency fixes, bitmap sharing for particles, and have Player code to allow a player to move around on screen

+++ Tartarus/src/org/ubc/tartarus/Point.java
@@ -1,9 +1,9 @@
 package org.ubc.tartarus;
 
 public class Point {
-	public int x, y; 
+	public float x, y; 
 	
-	public Point(int x, int y) {
+	public Point(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
